backend/pyproject.toml
[tool.poetry]
name = "todo-backend-api"
version = "0.1.0"
description = "Backend API for Todo Full-Stack Web Application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
sqlmodel = "^0.0.22"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.20"
psycopg2-binary = "^2.9.10"
alembic = "^1.13.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.23.8"
black = "^24.8.0"
flake8 = "^7.1.1"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ['E203', 'W503']  # Black compatibility
per-file-ignores = [
    '__init__.py:F401',  # Allow unused imports in __init__.py
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true